requirements.txt
discord.py
google_translate_py
pillow
pycountry
pynacl
youtube_dl